feat: enhance ticket management with new functionalities and refactor DAO connections
