+++ java/src/Main.java
@@ -1,24 +1,22 @@
-import java.awt.*;
 import java.sql.Connection;
 import java.sql.SQLException;
-
 import config.ConnectionFactory;
-import view.Menu;
+import view.MenuTicket;
 
 public class Main {
     public static void main(String[] args) {
-        Menu menu = new Menu();
-        menu.start();
 
-        // Check connection to DB
+        // Verificar conexión con Supabase antes de iniciar
         try (Connection conn = ConnectionFactory.getConnection()) {
-            System.out.println("Connected to Supabase successfully!");
-
+            System.out.println("✅ Connected to Supabase successfully!");
         } catch (SQLException e) {
-            System.err.println("Connection failed:");
+            System.err.println("❌ Connection failed:");
             e.printStackTrace();
-
+            return;
         }
-    }
 
+        // Iniciar menú principal
+        MenuTicket menuTicket = new MenuTicket();
+        menuTicket.start();
+    }
 }
