+++ java/src/dao/CommentDaoJDBC.java
@@ -1,6 +1,6 @@
 package dao;
 
-import config.DbConfig;
+import config.ConnectionFactory;
 import domain.Comentario;
 
 import java.sql.*;
@@ -12,7 +12,7 @@ public class CommentDaoJDBC implements CommentDao {
     @Override
     public void addComment(Comentario comentario) {
         String sql = "INSERT INTO comments (ticket_id, user_id, text, created_at) VALUES (?, ?, ?, ?)";
-        try (Connection conn = DbConfig.getConnection();
+        try (Connection conn = ConnectionFactory.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
 
             pstmt.setInt(1, comentario.getTicketId());
@@ -37,7 +37,7 @@ public class CommentDaoJDBC implements CommentDao {
     public List<Comentario> listByTicket(int ticketId) {
         List<Comentario> comentarios = new ArrayList<>();
         String sql = "SELECT * FROM comments WHERE ticket_id = ? ORDER BY created_at ASC";
-        try (Connection conn = DbConfig.getConnection();
+        try (Connection conn = ConnectionFactory.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
             pstmt.setInt(1, ticketId);
