+++ java/src/dao/TicketDao.java
@@ -1,11 +1,14 @@
 package dao;
 
 import domain.Ticket;
+
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 public interface TicketDao {
 
-    Ticket crear(Ticket ticket);
+    void save(Ticket ticket);
 
     Optional<Ticket> findById(int ticketId);
 
@@ -14,5 +17,25 @@ public interface TicketDao {
     void assign(int ticketId, int assigneeId);
 
     void changeState(int ticketId, int stateId);
+
+    List<Ticket> findAll();
+
+    List<Ticket> listByStateAndCategory(String stateName, String categoryName);
+
+    List<Ticket> listByAssignee(String assigneeUsername);
+
+    Map<String, Integer> topCategories(int limit);
+
+     int findCategoryIdByName(String name);
+
+     int findStateIdByName(String name);
+
+     int findUserIdByUsername(String username);
+
+     List<String> listAllCategories();
+
+     List<String> listAllStates();
+
+     List<String> listAllUsernames();
 }
